added Timer component
